Initial extraction of re-usable bits.
This generalizes the integration-test setup and moves all re-usable bits
to a new module, the `texasbbq`. This PR tracks extraction of that
module, eventual seperation into a package and re-integration of that
package into the integration-testing
